refactor: 소스코드 디렉토리 구조 개선 및 파일 위치 최적화
