+++ src/styles/components/Settings.css
